Extract mock editor; misc review feedback
Jira ticket: CAMS-526
